hardware/oobavd/requirements.txt
torch
torchvision
tqdm
pillow
pyusb
python-magic
numpy
fastai